Implement BitwiseIp struct
Now that this is a separate library from RemoteIp, I have the room to
stretch out and define modules & structs that make sense for the data
structures (rather than dumping everything into RemoteIp.Block). As a
first step, we have to handle parsing & encoding IP addresses as
integers upon which we'll perform bit arithmetic.