Add BitwiseIp.Mask
Unlike RemoteIp and its runtime algorithms, this module statically
compiles into pattern-matching functions to do all its logic. This
theoretically gives us faster parsing because we don't have to hop
through the hoops of String.to_integer/1, bitwise operations, etc.

This works because the domain is already rather small for all of the
possible masks. There are only 33 + 129 = 162 values, so having that
many pattern-matching clauses shouldn't be too taxing. I have yet to
benchmark this, though, so we'll have to test it to make sure.