Add BitwiseIp.Blocks
Per the discussion in ajvondrak/remote_ip#27, the Blocks.contain?/2
function will make it a bit easier to avoid the overhead of re-encoding
the IP repeatedly in the Enum.any?/2 loop. This even makes it
appropriate to expose a branch of the pattern that deals with
:inet-style tuples.

As long as I'm adding this module, the parsing functions also help sift
through lists of strings, either discarding errors or eagerly raising
them.