Teach BitwiseIp.Blocks to merge overlapping blocks
This is kind of a carnival trick I'm borrowing from the cider library
(https://hex.pm/packages/cider), but it makes enough sense. The driving
use case for BitwiseIp is maintaining a list of blocks and checking for
IP membership in any one of them. The fewer blocks you have, the less
work Blocks.contain?/2 has to do. To that end, we can remove any block
from the list that's wholly contained in another. The algorithm follows
as a simplistic iteration until fixed point, where each iteration tries
to greedily merge two blocks in the list.