Have Blocks.optimize/1 put widest blocks first
After collapsing all the blocks together, a further step we can take to
optimize Blocks.contain?/2 is to reorder them such that we're more
likely to get a positive result early. That is, when the mask is small,
the space covered by the block is large, so it's more likely that
Block.member?/2 will return true. The membership check is still the same
complexity regardless, but this might save a few iterations on the loop
through the list of blocks.