Factor out BitwiseIp.Block.size/1
Instead of decoding a bit string at runtime, this allows the Enumerable
count to be a constant lookup compiled directly into the module. I
haven't benchmarked this, but it's a similar idea to the whole
implementation of BitwiseIp.Mask.

The Enumerable tests already cover size/1 just fine, so I didn't add any
redundant cases.