Denormalize BitwiseIp.Block struct
This is a pretty significant refactoring. Instead of using the :prefix
field as its own %BitwiseIp{}, the protocol + address values are
inlined. My guess is that this is a more efficient layout because then
we don't box an extra map and instead just have one "flat" map of
integers/symbols. At any rate, the results are back to what we expect
(in line with RemoteIp):

  remote_ip          16.11
  bitwise_ip         16.10 - 1.00x slower +0.0488 ms
  cider              12.22 - 1.32x slower +19.78 ms
  inet_cidr           6.98 - 2.31x slower +81.14 ms
  cidr                1.65 - 9.75x slower +543.03 ms

With this change, I kind of reworked a bunch of code to leverage more
pattern matching and/or inline values, less dot notation (lessons
learned from member?/2). With all this going on, I found I needed
shorter names just to keep things single-line and not too awful. So I
shortened them to :proto and :addr, both in the Block and in the
BitwiseIp for consistency.