Remove remote_ip from benchmarks
This was for basic regression testing, but RemoteIp.Block isn't part of
its public interface, so it's misleading to keep it around in the
benchmarks.

I think it still edges out the implementation of BitwiseIp.Block just
because it's using tuples like {:v4, ...} instead of structs like
%BitwiseIp{proto: :v4, addr: ...}. But I'm not going to golf it down
that much. There's a limit to my squeezing blood from the stone.