Replace `use Bitwise` with `import Bitwise`
Note that `use Bitwise` was deprecated in Elixir 1.14, but `import
Bitwise` has been available since 1.0, so this should be fine for our
current compatibility against Elixir 1.12 and above. CI will let us know
if this is a bad assumption, but `mix test` passes without deprecation
warnings on Elixir 1.16 locally.

Fixes #2.