Update benchmark scripts
They also had the same `use Bitwise` -> `import Bitwise` needed for #2,
and their dependencies were out of date, so I went ahead and put some of
that polish on the scripts.

Running all of them again, it doesn't seem like the results have
materially changed since the last release of bitwise_ip, so I'm not
bothering to update the README or anything.